# frontend/src/lib/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_create_identification_identifications_post": {
        "properties": {
          "file": {
            "description": "Plant image to identify",
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_create_identification_identifications_post",
        "type": "object"
      },
      "CandidateEnrichmentStatus": {
        "properties": {
          "candidate_id": {
            "format": "uuid",
            "title": "Candidate Id",
            "type": "string"
          },
          "job": {
            "$ref": "#/components/schemas/JobStatusResponse"
          },
          "policy_version": {
            "minimum": 1.0,
            "title": "Policy Version",
            "type": "integer"
          }
        },
        "required": [
          "candidate_id",
          "policy_version",
          "job"
        ],
        "title": "CandidateEnrichmentStatus",
        "type": "object"
      },
      "ConfirmationResponse": {
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/TaxonomyCandidate"
          },
          "enrichment": {
            "$ref": "#/components/schemas/CandidateEnrichmentStatus"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "candidate",
          "enrichment"
        ],
        "title": "ConfirmationResponse",
        "type": "object"
      },
...
      "EnrichmentJobResult": {
        "additionalProperties": false,
        "properties": {
          "acquisition_avoided": {
            "default": false,
            "title": "Acquisition Avoided",
            "type": "boolean"
          },
          "covered_aspects": {
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "title": "Covered Aspects",
            "type": "array"
          },
          "covered_count": {
            "maximum": 32.0,
            "minimum": 0.0,
            "title": "Covered Count",
            "type": "integer"
          },
          "limitations": {
            "items": {
              "$ref": "#/components/schemas/EnrichmentLimitation"
            },
            "maxItems": 10,
            "title": "Limitations",
            "type": "array"
          },
          "missing_aspects": {
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "title": "Missing Aspects",
            "type": "array"
          },
          "missing_count": {
            "maximum": 32.0,
            "minimum": 0.0,
            "title": "Missing Count",
            "type": "integer"
          },
          "outcome": {
            "enum": [
              "complete",
              "partial"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "policy_version": {
            "minimum": 1.0,
            "title": "Policy Version",
            "type": "integer"
          }
        },
        "required": [
          "outcome",
          "policy_version",
          "covered_aspects",
          "missing_aspects",
          "covered_count",
          "missing_count"
        ],
        "title": "EnrichmentJobResult",
        "type": "object"
      },
      "EnrichmentLimitation": {
        "enum": [
          "missing_required_aspects",
          "safety_evidence_rejected"
        ],
        "title": "EnrichmentLimitation",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IdentificationResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/TaxonomyCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image": {
            "additionalProperties": true,
            "title": "Image",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "sad_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sad Path"
          },
          "status": {
            "$ref": "#/components/schemas/IdentificationStatus"
          }
        },
        "required": [
          "id",
          "status",
          "message",
          "image"
        ],
        "title": "IdentificationResponse",
        "type": "object"
      },
      "IdentificationStatus": {
        "enum": [
          "needs_confirmation",
          "retry_needed",
          "no_reliable_candidate"
        ],
        "title": "IdentificationStatus",
        "type": "string"
      },
      "JobFailureCategory": {
        "enum": [
          "invalid_payload",
          "unsupported_payload_version",
          "unknown_job_type",
          "database_transient",
          "provider_transient",
          "indexing_transient",
          "invariant_violation",
          "attempts_exhausted",
          "unexpected_error",
          "lease_expired",
          "lease_lost",
          "insufficient_evidence"
        ],
        "title": "JobFailureCategory",
        "type": "string"
      },
      "JobLimitation": {
        "enum": [
          "some_claims_failed",
          "indexing_deferred"
        ],
        "title": "JobLimitation",
        "type": "string"
      },
      "JobStatus": {
        "enum": [
          "pending",
          "processing",
          "complete",
          "partial",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobStatusResponse": {
        "properties": {
          "attempt_count": {
            "minimum": 0.0,
            "title": "Attempt Count",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "last_error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReadJobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "max_attempts": {
            "minimum": 1.0,
            "title": "Max Attempts",
            "type": "integer"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReadJobResult"
              },
              {
                "$ref": "#/components/schemas/EnrichmentJobResult"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "job_type",
          "status",
          "attempt_count",
          "max_attempts",
          "created_at",
          "updated_at"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
      "JobType": {
        "enum": [
          "ingest_validated_claims",
          "enrich_confirmed_plant"
        ],
        "title": "JobType",
        "type": "string"
      },
...
      "ReadJobError": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/JobFailureCategory"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "category"
        ],
        "title": "ReadJobError",
        "type": "object"
      },
      "ReadJobResult": {
        "properties": {
          "failed": {
            "default": 0,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Failed",
            "type": "integer"
          },
          "limitations": {
            "items": {
              "$ref": "#/components/schemas/JobLimitation"
            },
            "maxItems": 10,
            "title": "Limitations",
            "type": "array"
          },
          "partial": {
            "default": false,
            "title": "Partial",
            "type": "boolean"
          },
          "skipped": {
            "default": 0,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Skipped",
            "type": "integer"
          },
          "succeeded": {
            "default": 0,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Succeeded",
            "type": "integer"
          }
        },
        "title": "ReadJobResult",
        "type": "object"
      },
...
      "TaxonomyCandidate": {
        "properties": {
          "accepted_scientific_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted Scientific Name"
          },
          "binomial_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Binomial Name"
          },
          "common_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Name"
          },
          "confidence_label": {
            "title": "Confidence Label",
            "type": "string"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "family": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family"
          },
          "gbif_accepted_key": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gbif Accepted Key"
          },
          "gbif_key": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gbif Key"
          },
          "genus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genus"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "possible_match_copy": {
            "title": "Possible Match Copy",
            "type": "string"
          },
          "species": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Species"
          },
          "suggested_scientific_name": {
            "title": "Suggested Scientific Name",
            "type": "string"
          },
          "synonyms": {
            "items": {
              "type": "string"
            },
            "title": "Synonyms",
            "type": "array"
          },
          "taxonomic_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taxonomic Status"
          },
          "validation_status": {
            "$ref": "#/components/schemas/ValidationStatus"
          },
          "visible_traits": {
            "items": {
              "type": "string"
            },
            "title": "Visible Traits",
            "type": "array"
          }
        },
        "required": [
          "id",
          "suggested_scientific_name",
          "confidence_label",
          "possible_match_copy",
          "validation_status"
        ],
        "title": "TaxonomyCandidate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationStatus": {
        "enum": [
          "validated",
          "no_gbif_match"
        ],
        "title": "ValidationStatus",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "Fotosintesis AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/identifications": {
      "post": {
        "operationId": "create_identification_identifications_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "fotosintesis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fotosintesis Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_identification_identifications_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Identification",
        "tags": [
          "identifications"
        ]
      }
    },
    "/identifications/candidates/{candidate_id}/enrichment": {
      "get": {
        "operationId": "get_candidate_enrichment_identifications_candidates__candidate_id__enrichment_get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "fotosintesis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fotosintesis Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateEnrichmentStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Candidate Enrichment",
        "tags": [
          "identifications"
        ]
      }
    },
    "/identifications/{identification_id}/candidates/{candidate_id}/confirm": {
      "post": {
        "operationId": "confirm_candidate_identifications__identification_id__candidates__candidate_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "identification_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identification Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "fotosintesis_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fotosintesis Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Candidate",
        "tags": [
          "identifications"
        ]
      }
    },
...
  }
}